code refactoring
using abstract classes
